requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
yt-dlp==2023.11.16
aiohttp==3.9.1